refactor(gateway): merge llm-gateway + document-gateway into one control plane
Consolidate apps/llm-gateway and apps/document-gateway into a single
apps/gateway service that does both jobs behind one token-verify, while
preserving every security check both services did.

One Hono app registers, in correctness-critical order: /health → the
document routes (/v1/documents/search|fetch, scope-enforced) → the
catch-all Anthropic proxy (app.all("*"), path-allowlisted). Because the
document routes register before the catch-all, /v1/documents/* is handled
there and never falls through to the proxy. A single bearerClaims helper
runs verifyLlmToken on both route families.

Preserved verbatim: token verify (timing-safe, expiry, shape) + fail-closed
401 on both families; Anthropic path allowlist (/v1/messages,
/v1/messages/count_tokens) so it's not an open proxy; inbound request-header
allowlist + outbound hop-by-hop/encoding response-header stripping;
document scope enforcement (isKnownScope, scopeError, workspace pinning,
collection/document membership checks, parameterized SQL only); generic 502s
that leak no credentials/internals.

Env merges into one module requiring ANTHROPIC_API_KEY + DATABASE_URL +
LLM_TOKEN_SECRET, with a single GATEWAY_PORT (default 8080). chat-api
collapses LLM_GATEWAY_PUBLIC_URL + DOCUMENT_GATEWAY_PUBLIC_URL into one
GATEWAY_PUBLIC_URL; both turn fields (llm_base_url, doc_gateway_url) derive
from it. The daemon, mymemo-docs CLI, agent, and @mymemo/llm-token are
untouched.

Tradeoff (documented in AGENTS.md + docs): one process now holds BOTH
credentials with one egress identity reaching both Anthropic and the KB
Postgres — a wider blast radius, accepted as the cost of one deployable
control plane instead of two.

Tests: combined both suites under apps/gateway plus a routing-isolation
test that locks the registration order (document search hits the DB seam,
not the proxy; /v1/messages still proxies; other /v1/* still 404s; /health
answers tokenless). 32 pass, 6 integration tests skipped (need live PG).
compose, README, AGENTS.md, the architecture doc, and the gateway smoke
script updated; bun.lock regenerated for the new member set.
