+++ apps/chat-api/sandbox-template/template.ts
@@ -4,19 +4,11 @@ const WORKSPACE_ROOT = "/workspace";
 
 export const template = Template()
 	.fromNodeImage("24")
-	.aptInstall([
-		"curl",
-		"git",
-		"ripgrep",
-		"lsof",
-		"zstd",
-		"unzip",
-		"bubblewrap",
-	])
+	.aptInstall(["curl", "git", "ripgrep", "lsof", "zstd", "unzip", "bubblewrap"])
 	.runCmd("curl -fsSL https://bun.sh/install | bash")
 	.runCmd("ln -s /home/user/.bun/bin/bun /usr/local/bin/bun", { user: "root" })
 	// The agent's document access: a tiny CLI on PATH that forwards search/fetch
-	// to the document-gateway with the per-turn token (MYMEMO_DOC_* env vars set
+	// to the gateway's document routes with the per-turn token (MYMEMO_DOC_* env vars set
 	// by the daemon). On PATH under `/`, so it's visible inside the agent's bwrap
 	// (`--ro-bind / /`). Holds no credential itself.
 	.copy("mymemo-docs", "/usr/local/bin/mymemo-docs", {
