+++ apps/chat-api/src/config/env.ts
@@ -8,32 +8,20 @@ export const apiEnv = (() => {
 	invariant(Bun.env.E2B_API_KEY, "E2B_API_KEY is required");
 	invariant(Bun.env.DAEMON_AUTH_TOKEN, "DAEMON_AUTH_TOKEN is required");
 	invariant(Bun.env.LLM_TOKEN_SECRET, "LLM_TOKEN_SECRET is required");
-	invariant(
-		Bun.env.LLM_GATEWAY_PUBLIC_URL,
-		"LLM_GATEWAY_PUBLIC_URL is required",
-	);
-	invariant(
-		Bun.env.DOCUMENT_GATEWAY_PUBLIC_URL,
-		"DOCUMENT_GATEWAY_PUBLIC_URL is required",
-	);
+	invariant(Bun.env.GATEWAY_PUBLIC_URL, "GATEWAY_PUBLIC_URL is required");
 
 	return {
 		DAEMON_AUTH_TOKEN: Bun.env.DAEMON_AUTH_TOKEN,
 		// HMAC secret for the session tokens minted into each sandbox turn. Shared
-		// with llm-gateway and document-gateway, which both verify them (the
-		// document-gateway additionally enforces the token's signed scope).
+		// with the gateway, which verifies them and enforces the token's signed
+		// document scope.
 		LLM_TOKEN_SECRET: Bun.env.LLM_TOKEN_SECRET,
-		// Base URL the sandboxed agent points the Claude binary at
-		// (ANTHROPIC_BASE_URL). Must be reachable from inside the E2B sandbox.
-		// Trailing slash stripped so the binary's `${base}/v1/messages` never
-		// produces a double slash the gateway would have to normalize.
-		LLM_GATEWAY_PUBLIC_URL: Bun.env.LLM_GATEWAY_PUBLIC_URL.replace(/\/+$/, ""),
-		// Document gateway base URL the sandbox agent's `mymemo-docs` CLI calls.
-		// Must be reachable from inside the E2B sandbox. Trailing slash stripped.
-		DOCUMENT_GATEWAY_PUBLIC_URL: Bun.env.DOCUMENT_GATEWAY_PUBLIC_URL.replace(
-			/\/+$/,
-			"",
-		),
+		// Base URL of the merged gateway. The sandboxed agent points BOTH the Claude
+		// binary (ANTHROPIC_BASE_URL → /v1/messages) and the `mymemo-docs` CLI
+		// (MYMEMO_DOC_GATEWAY_URL → /v1/documents/*) at this one service. Must be
+		// reachable from inside the E2B sandbox. Trailing slash stripped so the
+		// binary's `${base}/v1/messages` never produces a double slash.
+		GATEWAY_PUBLIC_URL: Bun.env.GATEWAY_PUBLIC_URL.replace(/\/+$/, ""),
 		LOG_LEVEL: Bun.env.LOG_LEVEL || "info",
 		E2B_TEMPLATE: Bun.env.E2B_TEMPLATE || "sandbox-template-dev",
 	} as const;
