+++ apps/chat-api/src/test-setup.ts
@@ -4,7 +4,5 @@ Bun.env.E2B_API_KEY = Bun.env.E2B_API_KEY ?? "test-e2b-key";
 Bun.env.DAEMON_AUTH_TOKEN =
 	Bun.env.DAEMON_AUTH_TOKEN ?? "test-daemon-auth-token";
 Bun.env.LLM_TOKEN_SECRET = Bun.env.LLM_TOKEN_SECRET ?? "test-llm-token-secret";
-Bun.env.LLM_GATEWAY_PUBLIC_URL =
-	Bun.env.LLM_GATEWAY_PUBLIC_URL ?? "https://gateway.test";
-Bun.env.DOCUMENT_GATEWAY_PUBLIC_URL =
-	Bun.env.DOCUMENT_GATEWAY_PUBLIC_URL ?? "https://docs.test";
+Bun.env.GATEWAY_PUBLIC_URL =
+	Bun.env.GATEWAY_PUBLIC_URL ?? "https://gateway.test";
