+++ apps/gateway/package.json
@@ -1,5 +1,5 @@
 {
-	"name": "document-gateway",
+	"name": "gateway",
 	"version": "0.1.0",
 	"type": "module",
 	"scripts": {
