+++ apps/gateway/src/index.test.ts
@@ -1,15 +1,17 @@
-import { afterEach, beforeAll, describe, expect, it } from "bun:test";
+import { afterEach, beforeAll, describe, expect, it, spyOn } from "bun:test";
 import { type LlmTokenClaims, mintLlmToken } from "@mymemo/llm-token";
 import type { Db } from "./db";
 
+// The merged env validates ANTHROPIC_API_KEY + DATABASE_URL + LLM_TOKEN_SECRET at
+// module load, so set all three before importing ./index.
 const SECRET = "test-secret";
+Bun.env.ANTHROPIC_API_KEY = "test-anthropic-key";
+Bun.env.DATABASE_URL = "postgres://test@localhost/test";
+Bun.env.LLM_TOKEN_SECRET = SECRET;
 
 let app: typeof import("./index").app;
 let setDbForTests: typeof import("./index").setDbForTests;
-
 beforeAll(async () => {
-	Bun.env.LLM_TOKEN_SECRET = SECRET;
-	Bun.env.DATABASE_URL = "postgres://test@localhost/test";
 	({ app, setDbForTests } = await import("./index"));
 });
 
@@ -31,7 +33,7 @@ interface Call {
 	params: unknown[];
 }
 let calls: Call[] = [];
-let responder: (text: string, params: unknown[]) => unknown[];
+let responder: (text: string, params: unknown[]) => unknown[] = () => [];
 
 const fakeDb: Db = {
 	async query<T>(text: string, params: unknown[] = []): Promise<T[]> {
@@ -49,13 +51,119 @@ function kind(text: string): "search" | "resolveDoc" | "resolveColl" | "fetch" {
 const callOf = (k: ReturnType<typeof kind>) =>
 	calls.find((c) => kind(c.text) === k);
 
-beforeAll(() => setDbForTests(fakeDb));
-afterEach(() => {
-	calls = [];
-	responder = () => [];
+// ── LLM proxy (verbatim from the former llm-gateway suite) ──
+describe("gateway · llm proxy", () => {
+	const validToken = () =>
+		mintLlmToken(
+			{ userId: "u1", sandboxId: "sbx-1", requestId: "req-1" },
+			SECRET,
+		);
+
+	let fetchSpy: ReturnType<typeof spyOn> | undefined;
+	afterEach(() => fetchSpy?.mockRestore());
+
+	it("answers GET and HEAD /health without a token", async () => {
+		expect((await app.request("/health")).status).toBe(200);
+		expect((await app.request("/health", { method: "HEAD" })).status).toBe(200);
+	});
+
+	it("rejects requests with no bearer token", async () => {
+		const res = await app.request("/v1/messages", { method: "POST" });
+		expect(res.status).toBe(401);
+	});
+
+	it("rejects an invalid bearer token", async () => {
+		const res = await app.request("/v1/messages", {
+			method: "POST",
+			headers: { authorization: "Bearer not-a-real-token" },
+		});
+		expect(res.status).toBe(401);
+	});
+
+	it("injects x-api-key and forwards anthropic headers for a valid token", async () => {
+		fetchSpy = spyOn(globalThis, "fetch").mockResolvedValue(
+			new Response('{"ok":true}', {
+				status: 200,
+				headers: { "content-type": "application/json" },
+			}),
+		);
+
+		const res = await app.request("/v1/messages", {
+			method: "POST",
+			headers: {
+				authorization: `Bearer ${validToken()}`,
+				"anthropic-version": "2023-06-01",
+				"content-type": "application/json",
+			},
+			body: JSON.stringify({ model: "claude-sonnet-4-6", messages: [] }),
+		});
+
+		expect(res.status).toBe(200);
+		expect(fetchSpy).toHaveBeenCalledTimes(1);
+		const [url, init] = fetchSpy.mock.calls[0] as [string, RequestInit];
+		expect(url).toBe("https://api.anthropic.com/v1/messages");
+		const sent = new Headers(init.headers);
+		expect(sent.get("x-api-key")).toBe("test-anthropic-key");
+		expect(sent.get("anthropic-version")).toBe("2023-06-01");
+		expect(sent.has("authorization")).toBe(false);
+	});
+
+	it("404s a non-messages path even with a valid token (no upstream call)", async () => {
+		fetchSpy = spyOn(globalThis, "fetch").mockRejectedValue(
+			new Error("should not forward"),
+		);
+		const res = await app.request("/v1/files", {
+			method: "POST",
+			headers: { authorization: `Bearer ${validToken()}` },
+			body: "{}",
+		});
+		expect(res.status).toBe(404);
+		expect(fetchSpy).not.toHaveBeenCalled();
+	});
+
+	it("normalizes a trailing slash and still proxies /v1/messages", async () => {
+		fetchSpy = spyOn(globalThis, "fetch").mockResolvedValue(
+			new Response("{}", { status: 200 }),
+		);
+		const res = await app.request("/v1/messages/", {
+			method: "POST",
+			headers: {
+				authorization: `Bearer ${validToken()}`,
+				"content-type": "application/json",
+			},
+			body: "{}",
+		});
+		expect(res.status).toBe(200);
+		const [url] = fetchSpy.mock.calls[0] as [string, RequestInit];
+		expect(url).toBe("https://api.anthropic.com/v1/messages");
+	});
+
+	it("normalizes a double-slash path and still proxies /v1/messages", async () => {
+		fetchSpy = spyOn(globalThis, "fetch").mockResolvedValue(
+			new Response("{}", { status: 200 }),
+		);
+		const res = await app.request("//v1/messages", {
+			method: "POST",
+			headers: {
+				authorization: `Bearer ${validToken()}`,
+				"content-type": "application/json",
+			},
+			body: "{}",
+		});
+		expect(res.status).toBe(200);
+		const [url] = fetchSpy.mock.calls[0] as [string, RequestInit];
+		expect(url).toBe("https://api.anthropic.com/v1/messages");
+	});
 });
 
-describe("document-gateway (FTS / Postgres)", () => {
+// ── Document reader (verbatim from the former document-gateway suite) ──
+describe("gateway · document reader (FTS / Postgres)", () => {
+	beforeAll(() => setDbForTests(fakeDb));
+	afterEach(() => {
+		calls = [];
+		responder = () => [];
+	});
+
 	it("rejects requests without a valid token (no DB touched)", async () => {
 		const res = await app.request("/v1/documents/search", {
 			method: "POST",
@@ -329,5 +437,70 @@ describe("document-gateway (FTS / Postgres)", () => {
 			body: JSON.stringify({ documentId: "d1" }),
 		});
 		expect(res.status).toBe(502);
+	});
+});
+
+// ── Routing isolation: the merge must not let the two families collide ──
+// The document routes are registered before the LLM catch-all, so /v1/documents/*
+// is handled by the document reader and never proxied to Anthropic; /v1/messages
+// still proxies; other /v1/* still 404 on the proxy's path allowlist.
+describe("gateway · routing isolation", () => {
+	beforeAll(() => setDbForTests(fakeDb));
+	let fetchSpy: ReturnType<typeof spyOn> | undefined;
+	afterEach(() => {
+		fetchSpy?.mockRestore();
+		calls = [];
+		responder = () => [];
+	});
+
+	it("routes /v1/documents/search to the document handler, never the Anthropic proxy", async () => {
+		// fetch would only be called by the LLM proxy; assert it never is.
+		fetchSpy = spyOn(globalThis, "fetch").mockRejectedValue(
+			new Error("document request must not reach the Anthropic proxy"),
+		);
+		responder = () => []; // global search → empty result set
+		const res = await app.request("/v1/documents/search", {
+			method: "POST",
+			headers: headers(token({ scope: "global" })),
+			body: JSON.stringify({ query: "x" }),
+		});
+		expect(res.status).toBe(200);
+		expect(await res.json()).toEqual({ documents: [] });
+		expect(fetchSpy).not.toHaveBeenCalled();
+		// It hit the DB seam — i.e. the document handler ran.
+		expect(callOf("search")).toBeDefined();
+	});
+
+	it("still proxies /v1/messages to Anthropic with a valid token", async () => {
+		fetchSpy = spyOn(globalThis, "fetch").mockResolvedValue(
+			new Response("{}", { status: 200 }),
+		);
+		const res = await app.request("/v1/messages", {
+			method: "POST",
+			headers: headers(token({ scope: "global" })),
+			body: "{}",
+		});
+		expect(res.status).toBe(200);
+		expect(fetchSpy).toHaveBeenCalledTimes(1);
+		const [url] = fetchSpy.mock.calls[0] as [string, RequestInit];
+		expect(url).toBe("https://api.anthropic.com/v1/messages");
+	});
+
+	it("still 404s another /v1/* path (e.g. /v1/files) with a valid token", async () => {
+		fetchSpy = spyOn(globalThis, "fetch").mockRejectedValue(
+			new Error("should not forward"),
+		);
+		const res = await app.request("/v1/files", {
+			method: "POST",
+			headers: headers(token({ scope: "global" })),
+			body: "{}",
+		});
+		expect(res.status).toBe(404);
+		expect(fetchSpy).not.toHaveBeenCalled();
+	});
+
+	it("answers /health on GET and HEAD without a token", async () => {
+		expect((await app.request("/health")).status).toBe(200);
+		expect((await app.request("/health", { method: "HEAD" })).status).toBe(200);
 	});
 });
