+++ apps/gateway/src/queries.ts
