--- apps/llm-gateway/biome.json
@@ -1,4 +1,0 @@
-{
-	"root": false,
-	"extends": "//"
-}
