--- apps/llm-gateway/package.json
@@ -1,19 +1,0 @@
-{
-	"name": "llm-gateway",
-	"version": "0.1.0",
-	"type": "module",
-	"scripts": {
-		"dev": "bun run --hot src/index.ts",
-		"format": "biome format --write",
-		"lint": "biome lint --write",
-		"check": "biome check --write"
-	},
-	"dependencies": {
-		"@mymemo/llm-token": "workspace:*",
-		"hono": "^4.10.1",
-		"tiny-invariant": "^1.3.3"
-	},
-	"devDependencies": {
-		"@types/bun": "^1.3.11"
-	}
-}
